Java programming example: Pet Matchmaking app
The aim of this project is to provide a system for pet owners to find nearby mates for their pets. To use the system, titled Pawnder, a user inputs their own credentials as well as their pet's species, breed, age, neighborhood, and name. The system responds by confirming if.a matching breed is available and then giving a list of potential matches based on the system's library. The user then selects one of these options to make a match. An email is sent to the pet owner of the match to inform them of the event. The focus of this system is to filter through the available pets in the system's library and present ideal matches to the user based on their input credentials