Python example: Guided Laboratory process
The aim of this project was to use python to guide a laboratory technician through the steps of performing a process called buffer exchange. The program asks the user for input to gain details on a lab sample, and then uses this information to make calculations needed to perform the lab procedure.